cli: add option to ignore selection event on error

+++ src/pipe_io.rs
@@ -12,7 +12,8 @@ use filedescriptor::FileDescriptor;
 pub(crate) fn read_with_timeout(
     mime_types_with_files: Vec<(Box<str>, FileDescriptor)>,
     timeout: Duration,
-    return_error_on_timeout: bool,
+    return_err_on_timeout: bool,
+    return_err_on_error: bool,
     limit_bytes: u64,
     already_used_bytes: u64,
 ) -> Result<HashMap<Box<str>, Result<Box<[u8]>, Rc<ReadWithTimeoutError>>>, ReadWithTimeoutError> {
@@ -34,24 +35,29 @@ pub(crate) fn read_with_timeout(
         data: Vec<u8>,
     }
 
-    let mut remaining_fd_to_data: HashMap<i32, FdData> = mime_types_with_files
-        .into_iter()
-        .filter_map(|(mime_type, mut file)| {
-            if let Err(err) = file.set_non_blocking(true) {
-                mime_type_to_data.insert(mime_type, Err(Rc::new(ReadWithTimeoutError::NonBlockingIOMode(err))));
-                return None;
+    let mut remaining_fd_to_data: HashMap<i32, FdData> = HashMap::with_capacity(mime_types_with_files.len());
+
+    for (mime_type, mut file) in mime_types_with_files {
+        if let Err(err) = file.set_non_blocking(true) {
+            let final_err = ReadWithTimeoutError::NonBlockingIOMode(err);
+
+            if return_err_on_error {
+                return Err(final_err);
             }
 
-            let raw_fd = file.as_raw_fd();
-            let fd_data = FdData {
-                mime_type,
-                file,
-                data: Vec::with_capacity(32),
-            };
+            mime_type_to_data.insert(mime_type, Err(Rc::new(final_err)));
+            continue;
+        }
 
-            Some((raw_fd, fd_data))
-        })
-        .collect();
+        let raw_fd = file.as_raw_fd();
+        let fd_data = FdData {
+            mime_type,
+            file,
+            data: Vec::with_capacity(32),
+        };
+
+        remaining_fd_to_data.insert(raw_fd, fd_data);
+    }
 
     // Create a list of remaining file descriptors to poll.
     let mut remaining_pfds: Vec<libc::pollfd> = remaining_fd_to_data
@@ -65,6 +71,7 @@ pub(crate) fn read_with_timeout(
 
     let mut buf = [0u8; 8192];
     let end_time = Instant::now() + timeout;
+    let mut read_error: Option<ReadWithTimeoutError> = None;
 
     // Convenience function to error out the remaining mime types.
     let error_out_remaining_mime_types =
@@ -87,41 +94,39 @@ pub(crate) fn read_with_timeout(
             .unwrap();
 
         if remaining_time == 0 {
-            if return_error_on_timeout {
-                return Err(ReadWithTimeoutError::Timeout);
+            let final_err = ReadWithTimeoutError::Timeout;
+
+            if return_err_on_timeout {
+                return Err(final_err);
             }
 
-            error_out_remaining_mime_types(
-                &mut mime_type_to_data,
-                remaining_fd_to_data,
-                ReadWithTimeoutError::Timeout,
-            );
+            error_out_remaining_mime_types(&mut mime_type_to_data, remaining_fd_to_data, final_err);
             break 'outer;
         }
 
         // Wait for changes of readability with timeout.
         match unsafe { libc::poll(remaining_pfds.as_mut_ptr(), remaining_pfds.len() as u64, remaining_time) } {
             0 => {
-                if return_error_on_timeout {
-                    return Err(ReadWithTimeoutError::Timeout);
+                let final_err = ReadWithTimeoutError::Timeout;
+
+                if return_err_on_timeout {
+                    return Err(final_err);
                 }
 
                 // Timeout occurred, error out the remaining mime types.
-                error_out_remaining_mime_types(
-                    &mut mime_type_to_data,
-                    remaining_fd_to_data,
-                    ReadWithTimeoutError::Timeout,
-                );
+                error_out_remaining_mime_types(&mut mime_type_to_data, remaining_fd_to_data, final_err);
                 break 'outer;
             }
             -1 => {
                 // Some other error occurred, error out the remaining mime types.
                 let err = std::io::Error::last_os_error();
-                error_out_remaining_mime_types(
-                    &mut mime_type_to_data,
-                    remaining_fd_to_data,
-                    ReadWithTimeoutError::Poll(err),
-                );
+                let final_err = ReadWithTimeoutError::Poll(err);
+
+                if return_err_on_error {
+                    return Err(final_err);
+                }
+
+                error_out_remaining_mime_types(&mut mime_type_to_data, remaining_fd_to_data, final_err);
                 break 'outer;
             }
             1.. => {
@@ -129,6 +134,11 @@ pub(crate) fn read_with_timeout(
                 remaining_pfds.retain_mut(|pfd| {
                     if exceeded_size_limit {
                         // Just ignore reading the data when the size limit has been exceeded.
+                        return true;
+                    }
+
+                    if read_error.is_some() {
+                        // Just ignore reading the data when a read error has occurred.
                         return true;
                     }
 
@@ -152,6 +162,7 @@ pub(crate) fn read_with_timeout(
                                         let owned_mime_type = removed_entry.1.mime_type;
                                         let data = removed_entry.1.data.into_boxed_slice();
                                         mime_type_to_data.insert(owned_mime_type, Ok(data));
+
                                         // Also remove this pfd from the remaining pfds.
                                         return false;
                                     }
@@ -175,13 +186,21 @@ pub(crate) fn read_with_timeout(
                                     }
                                     Err(err) => {
                                         // Some error occurred.
+                                        let final_err = ReadWithTimeoutError::Read(err);
+
+                                        if return_err_on_error {
+                                            read_error = Some(final_err);
+                                            return true;
+                                        }
+
                                         // Remove entry from data map and insert error to result.
                                         let removed_entry = entry.remove_entry();
                                         let owned_mime_type = removed_entry.1.mime_type;
-                                        mime_type_to_data
-                                            .insert(owned_mime_type, Err(Rc::new(ReadWithTimeoutError::Read(err))));
+                                        mime_type_to_data.insert(owned_mime_type, Err(Rc::new(final_err)));
+
                                         // Size limit: subtract this size from the current total size.
                                         current_size -= removed_entry.1.data.len() as u64;
+
                                         // Also remove this pfd from the remaining pfds.
                                         return false;
                                     }
@@ -201,14 +220,20 @@ pub(crate) fn read_with_timeout(
                 if exceeded_size_limit {
                     return Err(ReadWithTimeoutError::SizeLimitExceeded);
                 }
+
+                if let Some(read_error) = read_error {
+                    return Err(read_error);
+                }
             }
             return_value => {
                 // Invalid return value, error out the remaining mime types.
-                error_out_remaining_mime_types(
-                    &mut mime_type_to_data,
-                    remaining_fd_to_data,
-                    ReadWithTimeoutError::PollInvalidReturnValue(return_value),
-                );
+                let final_err = ReadWithTimeoutError::PollInvalidReturnValue(return_value);
+
+                if return_err_on_error {
+                    return Err(final_err);
+                }
+
+                error_out_remaining_mime_types(&mut mime_type_to_data, remaining_fd_to_data, final_err);
                 break 'outer;
             }
         }
